수업 테이블에 status추가, 수업삭제, 수업에 참여한 멤버 삭제 가능(수정중), 수업 관련 데이터 관련 클래스 패키지 변경(서버 구동 에러시 다시빌드 바람)

+++ ngiiedu/src/main/java/kr/go/ngii/edu/main/courses/course/service/CourseMemberService.java
@@ -152,7 +152,28 @@ public class CourseMemberService {
 	 * @return
 	 */
 	public boolean leave(int courseId, int userId) {
-
+		
+		if (courseMemberMapper.exists(courseId, userId)) {
+			courseMemberMapper.deleteByCourseIdAndUserId(courseId, userId);
+			return true;
+		} else {
+			return false;
+		}
+	}
+	
+	/**
+	 * 수업에서 멤버 삭제하기
+	 * 
+	 * @param courseId
+	 * @param userId
+	 * @return
+	 */
+	public boolean leave(int courseId, int userId, String password) {
+		// 패스워드 체크필요함.
+		if (false) {
+			throw new RuntimeException(ErrorMessage.PASSWORD_AUTHENTICATION_FAILED);
+		}
+		
 		if (courseMemberMapper.exists(courseId, userId)) {
 			courseMemberMapper.deleteByCourseIdAndUserId(courseId, userId);
 			return true;
